Started CI/CD setup; Included tox, mypy, flake8, pytest

+++ src/DataManager/utils/timehandler.py
@@ -14,8 +14,8 @@ class TimeHandler:
 
     @staticmethod
     def get_string_from_datetime64(inputDatetime64: np.datetime64) -> str:
-        return datetime.utcfromtimestamp((inputDatetime64 - np.datetime64('1970-01-01T00:00:00Z'))
-                                         / np.timedelta64(1, 's')).strftime('%Y-%m-%d %H:%M:%S')
+        return datetime.utcfromtimestamp(float((inputDatetime64 - np.datetime64('1970-01-01T00:00:00Z'))
+                                         / np.timedelta64(1, 's'))).strftime('%Y-%m-%d %H:%M:%S')
 
     @staticmethod
     def get_datetime64_from_string(inputString: str) -> np.datetime64:
